+++ src/main/java/com/enyoi/apimesaayuda/aplicacion/services/impl/TiposSolicitudService.java
@@ -1,9 +1,8 @@
 package com.enyoi.apimesaayuda.aplicacion.services.impl;
 
 import com.enyoi.apimesaayuda.aplicacion.dtos.TiposSolicitudDTO;
-import com.enyoi.apimesaayuda.aplicacion.entities.EstadosSolicitud;
 import com.enyoi.apimesaayuda.aplicacion.entities.TiposSolicitud;
-import com.enyoi.apimesaayuda.aplicacion.payloads.requests.TiposSolicitudRequests;
+import com.enyoi.apimesaayuda.aplicacion.payloads.requests.TiposSolicitudRequest;
 import com.enyoi.apimesaayuda.aplicacion.repositories.TiposSolicitudRepository;
 import com.enyoi.apimesaayuda.aplicacion.services.ITiposSolicitudService;
 import com.enyoi.apimesaayuda.base.exceptions.AlreadyExists;
@@ -13,7 +12,6 @@ import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.lang.reflect.Type;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -22,6 +20,7 @@ import java.util.Optional;
 @RequiredArgsConstructor
 public class TiposSolicitudService implements ITiposSolicitudService {
     private final ModelMapper modelMapper;
+    private static final String NOEXISTENDATOS = "No esxisten datos";
     private final TiposSolicitudRepository tiposSolicitudRepository;
 
 
@@ -43,10 +42,9 @@ public class TiposSolicitudService implements ITiposSolicitudService {
         TiposSolicitud tiposSolicitud;
         if (tiposSolicitudOptional.isPresent()) {
             tiposSolicitud = tiposSolicitudOptional.get();
-            TiposSolicitudDTO tiposSolicitudDTO = modelMapper.map(tiposSolicitud, TiposSolicitudDTO.class);
-            return tiposSolicitudDTO;
+            return modelMapper.map(tiposSolicitud, TiposSolicitudDTO.class);
         } else {
-            throw new NotDataFound("Tipo solicitud no existe");
+            throw new NotDataFound(NOEXISTENDATOS);
         }
 
     }
@@ -57,10 +55,9 @@ public class TiposSolicitudService implements ITiposSolicitudService {
         TiposSolicitud tiposSolicitud;
         if (tiposSolicitudOptional.isPresent()) {
             tiposSolicitud = tiposSolicitudOptional.get();
-            TiposSolicitudDTO tiposSolicitudDTO = modelMapper.map(tiposSolicitud, TiposSolicitudDTO.class);
-            return tiposSolicitudDTO;
+            return modelMapper.map(tiposSolicitud, TiposSolicitudDTO.class);
         } else {
-            throw new NotDataFound("Tipo solicitud no existe");
+            throw new NotDataFound(NOEXISTENDATOS);
         }
     }
 
@@ -73,24 +70,22 @@ public class TiposSolicitudService implements ITiposSolicitudService {
         } else {
             TiposSolicitud tiposSolicitud = new TiposSolicitud();
             tiposSolicitud.setTipoSolicitud(tipoSolicitud);
-            TiposSolicitudDTO tiposSolicitudDTO = modelMapper.map(tiposSolicitudRepository.save(tiposSolicitud), TiposSolicitudDTO.class);
-            return tiposSolicitudDTO;
+            return modelMapper.map(tiposSolicitudRepository.save(tiposSolicitud), TiposSolicitudDTO.class);
         }
 
     }
 
     @Override
-    public TiposSolicitudDTO update(TiposSolicitudRequests tiposSolicitudRequests) {
+    public TiposSolicitudDTO update(TiposSolicitudRequest tiposSolicitudRequests) {
         Optional<TiposSolicitud> tiposSolicitudOptional = tiposSolicitudRepository.findById(tiposSolicitudRequests.getTipoSolicitudId());
         if (tiposSolicitudOptional.isPresent()) {
             TiposSolicitud tipoSolicitudGuardar = tiposSolicitudOptional.get();
             tipoSolicitudGuardar.setTipoSolicitud(tiposSolicitudRequests.getTipoSolicitud() != null
                     ? tiposSolicitudRequests.getTipoSolicitud()
                     : tipoSolicitudGuardar.getTipoSolicitud());
-            TiposSolicitudDTO tiposSolicitudDTO = modelMapper.map(tiposSolicitudRepository.save(tipoSolicitudGuardar), TiposSolicitudDTO.class);
-            return tiposSolicitudDTO;
+            return modelMapper.map(tiposSolicitudRepository.save(tipoSolicitudGuardar), TiposSolicitudDTO.class);
         } else {
-            throw new NotDataFound("Tipo solicitud no existe");
+            throw new NotDataFound(NOEXISTENDATOS);
         }
     }
 
@@ -99,7 +94,12 @@ public class TiposSolicitudService implements ITiposSolicitudService {
     public String delete(Long id) {
         Optional<TiposSolicitud> tiposSolicitudOptional = Optional.ofNullable(tiposSolicitudRepository.findById(id)
                 .orElseThrow(() -> new NotDataFound("No existe el Tipo solicitud: " + id)));
-        tiposSolicitudRepository.deleteById(id);
-        return tiposSolicitudOptional.get() + "Tipo solicitud eliminado con Exito";
+        if (tiposSolicitudOptional.isPresent()){
+            tiposSolicitudRepository.deleteById(id);
+            return modelMapper.map(tiposSolicitudOptional.get(), TiposSolicitudDTO.class).getTipoSolicitud() + "Tipo solicitud eliminado con Exito";
+        }else {
+            throw new NotDataFound(NOEXISTENDATOS);
+        }
+
     }
 }
